+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 1.0.2 July 27, 2023
+- refactored plugin to OOP with PHP class loader
+- modified the Vue component to check for the site domain exposed by the Block Theme as part of data fetch
+- separated Vue and general scripts/styles into their own hook class files
+- minor logging and styles cleanup
+
 ## 1.0.1 July 25, 2023
 - Large update for "CleanBC Post Filter" block (added to Theme category in block inserter) includes:
 - funcitonal post, page or custom post type filtering